+++ include/raknet/StringTable.h
@@ -68,14 +68,14 @@ namespace RakNet
         /// \param[in] input Pointer to an ASCII string
         /// \param[in] maxCharsToWrite The size of \a input
         /// \param[out] output The bitstream to write the compressed string to
-        void EncodeString( const char *input, int maxCharsToWrite, RakNet::BitStream *output );
+        void EncodeString(const char *input, size_t maxCharsToWrite, RakNet::BitStream *output);
 
         /// Writes input to output, uncompressed.  Takes care of the null terminator for you.
         /// Relies on the StringCompressor class, which is automatically reference counted in the constructor and destructor in RakPeer.  You can call the reference counting functions yourself if you wish too.
         /// \param[out] output A block of bytes to receive the output
         /// \param[in] maxCharsToWrite Size, in bytes, of \a output .  A NULL terminator will always be appended to the output string.  If the maxCharsToWrite is not large enough, the string will be truncated.
         /// \param[in] input The bitstream containing the compressed string
-        bool DecodeString( char *output, int maxCharsToWrite, RakNet::BitStream *input );
+        bool DecodeString(char *output, size_t maxCharsToWrite, RakNet::BitStream *input);
 
         /// Private Constructor
         StringTable();
